Refactored overmind and made changes for members
